Add interactive mode and refactor command processing logic
- Make `command_in_natural_language` optional in CLI arguments
- Extract command processing logic into separate helper module
- Add comprehensive docs for `next_step` in SemiAutonomousCommandLineAgent
- Implement interactive mode with continuous user input
- Maintain context between commands in interactive mode
- Add exit option ('e') in interactive mode
